+++ packages/webamp/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Upcoming [UNRELEASED] (`webamp@next`)
 
-_No changes yet._
+### Improvements
+
+- Add new config option `enableMediaSession` to allow Webamp to connect to the browser's Media Session API. This enables OS/hardware level media controls like play/pause/next/previous.
 
 ## 2.1.2 [CURRENT]